+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[8.1].define(version: 2026_09_26_120000) do
+ActiveRecord::Schema[8.1].define(version: 2026_09_26_130000) do
   create_table "active_storage_attachments", force: :cascade do |t|
     t.bigint "blob_id", null: false
     t.datetime "created_at", null: false
@@ -239,6 +239,7 @@ ActiveRecord::Schema[8.1].define(version: 2026_09_26_120000) do
     t.datetime "created_at", null: false
     t.datetime "elevated_at"
     t.string "ip_address"
+    t.datetime "last_active_at"
     t.datetime "updated_at", null: false
     t.string "user_agent"
     t.integer "user_id", null: false